hiver_events::condition

Trait ConditionPropertyProvider

pub trait ConditionPropertyProvider {
    // Required method
    fn get_property(&self, path: &str) -> Option<String>;
}

Trait for events that can provide property values by path 可以按路径提供属性值的事件trait

Events implementing this trait can be filtered using PropertyCondition. The property path uses dot notation (e.g., “user.name”).

实现此trait的事件可以使用 PropertyCondition 进行过滤。 属性路径使用点号表示法（例如 “user.name”）。

Required Methods

fn get_property(&self, path: &str) -> Option<String>

Get a property value by its path 按路径获取属性值

Returns Some(value) if the property exists, None otherwise. 如果属性存在则返回 Some(value)，否则返回 None。